Added h2 database and react front end

+++ src/main/java/com/nidalas/wealthrating/entity/FinancialInfo.java
@@ -4,6 +4,6 @@ import lombok.Data;
 
 @Data
 public class FinancialInfo {
-    private float cash;
+    private double cash;
     private int numberOfAssets;
 }
