+++ src/main/java/com/nidalas/wealthrating/entity/WealthThresholdEntity.java
@@ -16,5 +16,5 @@ public class WealthThresholdEntity {
     private Integer id;
 
     @Column(name = "threshold")
-    private float 	threshold;
+    private double 	threshold;
 }
